complete version 1

+++ word_of_mouth/lib/core/components/products/product_card.dart
@@ -17,8 +17,8 @@ class ProductCard extends StatelessWidget {
     required this.press,
   });
   final String image, title;
-  final double? price;
-  final double? priceAfetDiscount;
+  final int? price;
+  final int? priceAfetDiscount;
   final int? dicountpercent;
   final VoidCallback press;
